requirements.txt
ddt==1.2.1
gitdb2==2.0.5
GitPython==3.0.1
smmap2==2.0.5
